General: Move ChangeTimes out of Geocoding
