+++ infra/src/main/kotlin/fi/fta/geoviite/infra/geocoding/Geocoding.kt
@@ -12,7 +12,6 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.math.BigDecimal
 import java.math.RoundingMode
-import java.time.Instant
 import kotlin.math.PI
 
 data class AddressPoint(val point: LayoutPoint, val address: TrackMeter, val distance: Double) {
@@ -34,13 +33,6 @@ data class AlignmentAddresses(
 data class AlignmentStartAndEnd(
     val start: AddressPoint?,
     val end: AddressPoint?,
-)
-
-data class ChangeTimes(
-    val created: Instant,
-    val changed: Instant,
-    val officialChanged: Instant?,
-    val draftChanged: Instant?,
 )
 
 data class ProjectionLine(val address: TrackMeter, val projection: Line)
