+++ infra/src/main/kotlin/fi/fta/geoviite/infra/util/DaoBase.kt
@@ -5,7 +5,7 @@ import fi.fta.geoviite.infra.common.IntId
 import fi.fta.geoviite.infra.error.NoSuchEntityException
 import fi.fta.geoviite.infra.logging.AccessType.VERSION_FETCH
 import fi.fta.geoviite.infra.logging.daoAccess
-import fi.fta.geoviite.infra.geocoding.ChangeTimes
+import fi.fta.geoviite.infra.tracklayout.ChangeTimes
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
